requirements.txt
streamlit>=1.32.0
pymupdf>=1.23.0
google-genai>=2.0.0
openai>=1.50.0
pandas>=2.0.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
numpy>=1.26.0
scipy>=1.12.0